rabbitmq/other_repos/03/requirements.txt
certifi==2025.1.31; python_version >= '3.6'
charset-normalizer==3.4.1; python_version >= '3.7'
idna==3.10; python_version >= '3.6'
pika==1.3.2; python_version >= '3.7'
python-dotenv==1.0.1; python_version >= '3.8'
requests==2.32.3; python_version >= '3.8'
urllib3==2.3.0; python_version >= '3.9'